feat:removeAllAppointmentsIdFromSchedules

+++ packages/api/src/appointments/appointments.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common'
+import { Inject, Injectable, forwardRef } from '@nestjs/common'
 import { CreateAppointmentInput } from './dto/create-appointment.input'
 import { UpdateAppointmentInput } from './dto/update-appointment.input'
 import { InjectRepository } from '@nestjs/typeorm'
@@ -11,12 +11,20 @@ import {
   filterAppointments,
   orderAppointments,
 } from 'src/helpers/appointmentsFunctions'
+import { SchedulesService } from 'src/schedules/schedules.service'
+import { UsersService } from 'src/users/users.service'
+import { LocationsService } from 'src/locations/locations.service'
 
 @Injectable()
 export class AppointmentsService {
   constructor(
     @InjectRepository(Appointment)
     private readonly appointmentRepository: Repository<Appointment>,
+    @Inject(forwardRef(() => SchedulesService))
+    private readonly scheduleService: SchedulesService,
+    @Inject(forwardRef(() => UsersService))
+    private readonly usersService: UsersService,
+    private readonly locationsService: LocationsService,
   ) {}
 
   async findAll(
@@ -48,8 +56,10 @@ export class AppointmentsService {
     createAppointmentInput: CreateAppointmentInput,
   ): Promise<Appointment> {
     const a = new Appointment()
-    a.userId = createAppointmentInput.userId
-    a.location = createAppointmentInput.location
+    a.user = await this.usersService.findOne(createAppointmentInput.userId)
+    a.location = await this.locationsService.findOne(
+      createAppointmentInput.locationId,
+    )
     a.type = createAppointmentInput.type
     a.endProposedDate = createAppointmentInput.endProposedDate
     a.startProposedDate = createAppointmentInput.startProposedDate
@@ -61,6 +71,7 @@ export class AppointmentsService {
     return this.appointmentRepository.save(a)
   }
 
+  // TODO: update user & locations
   async update(
     id: ObjectId,
     updateAppointmentInput: UpdateAppointmentInput,
@@ -75,16 +86,42 @@ export class AppointmentsService {
     return this.findOne(id.toString())
   }
 
+  // remove appointment only if not done & remove appointment id from schedules
   async remove(id: string): Promise<string> {
-    await this.findOne(id.toString())
+    const appointment = await this.findOne(id.toString())
+
+    if (appointment.isDone)
+      throw new GraphQLError('Cannot delete an appointment that is done')
+
+    // remove all appointments id from schedules & delete schedule if empty appointmentIds
+    await this.scheduleService.removeAllAppointmentsIdFromSchedules(id)
 
     await this.appointmentRepository.delete(id)
 
-    // TODO: delete the schedule that has this appointment id & finalDate => today
-    // Can only delete this appointment if IsDone = false
-
     // return id if delete was successful
     return id
+  }
+
+  // remove all appointments by userId only if not done
+  async removeAllByUserId(userId: string): Promise<string[]> {
+    const user = await this.usersService.findOne(userId)
+
+    const appointments = await this.appointmentRepository.find({
+      where: {
+        user: { uid: user.uid },
+        isDone: false,
+      },
+    })
+
+    if (appointments.length === 0) return []
+
+    const ids = appointments.map(appointment => appointment.id.toString())
+
+    for (const id of ids) {
+      await this.remove(id)
+    }
+
+    return ids
   }
 
   // Seeding functions
